+++ lib/twilio-ruby/rest/studio/v1/flow.rb
@@ -6,30 +6,27 @@
 
 module Twilio
   module REST
-    class Video < Domain
+    class Studio < Domain
       class V1 < Version
-        class RecordingList < ListResource
+        ##
+        # PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
+        class FlowList < ListResource
           ##
-          # Initialize the RecordingList
+          # Initialize the FlowList
           # @param [Version] version Version that contains the resource
-          # @return [RecordingList] RecordingList
+          # @return [FlowList] FlowList
           def initialize(version)
             super(version)
 
             # Path Solution
             @solution = {}
-            @uri = "/Recordings"
+            @uri = "/Flows"
           end
 
           ##
-          # Lists RecordingInstance records from the API as a list.
+          # Lists FlowInstance records from the API as a list.
           # Unlike stream(), this operation is eager and will load `limit` records into
           # memory before returning.
-          # @param [recording.Status] status The status
-          # @param [String] source_sid The source_sid
-          # @param [String] grouping_sid The grouping_sid
-          # @param [Time] date_created_after The date_created_after
-          # @param [Time] date_created_before The date_created_before
           # @param [Integer] limit Upper limit for the number of records to return. stream()
           #    guarantees to never return more than limit.  Default is no limit
           # @param [Integer] page_size Number of records to fetch per request, when
@@ -37,27 +34,14 @@ module Twilio
           #    but a limit is defined, stream() will attempt to read the limit with the most
           #    efficient page size, i.e. min(limit, 1000)
           # @return [Array] Array of up to limit results
-          def list(status: :unset, source_sid: :unset, grouping_sid: :unset, date_created_after: :unset, date_created_before: :unset, limit: nil, page_size: nil)
-            self.stream(
-                status: status,
-                source_sid: source_sid,
-                grouping_sid: grouping_sid,
-                date_created_after: date_created_after,
-                date_created_before: date_created_before,
-                limit: limit,
-                page_size: page_size
-            ).entries
+          def list(limit: nil, page_size: nil)
+            self.stream(limit: limit, page_size: page_size).entries
           end
 
           ##
-          # Streams RecordingInstance records from the API as an Enumerable.
+          # Streams FlowInstance records from the API as an Enumerable.
           # This operation lazily loads records as efficiently as possible until the limit
           # is reached.
-          # @param [recording.Status] status The status
-          # @param [String] source_sid The source_sid
-          # @param [String] grouping_sid The grouping_sid
-          # @param [Time] date_created_after The date_created_after
-          # @param [Time] date_created_before The date_created_before
           # @param [Integer] limit Upper limit for the number of records to return. stream()
           #    guarantees to never return more than limit. Default is no limit.
           # @param [Integer] page_size Number of records to fetch per request, when
@@ -65,23 +49,16 @@ module Twilio
           #    but a limit is defined, stream() will attempt to read the limit with the most
           #    efficient page size, i.e. min(limit, 1000)
           # @return [Enumerable] Enumerable that will yield up to limit results
-          def stream(status: :unset, source_sid: :unset, grouping_sid: :unset, date_created_after: :unset, date_created_before: :unset, limit: nil, page_size: nil)
+          def stream(limit: nil, page_size: nil)
             limits = @version.read_limits(limit, page_size)
 
-            page = self.page(
-                status: status,
-                source_sid: source_sid,
-                grouping_sid: grouping_sid,
-                date_created_after: date_created_after,
-                date_created_before: date_created_before,
-                page_size: limits[:page_size],
-            )
+            page = self.page(page_size: limits[:page_size], )
 
             @version.stream(page, limit: limits[:limit], page_limit: limits[:page_limit])
           end
 
           ##
-          # When passed a block, yields RecordingInstance records from the API.
+          # When passed a block, yields FlowInstance records from the API.
           # This operation lazily loads records as efficiently as possible until the limit
           # is reached.
           def each
@@ -95,24 +72,14 @@ module Twilio
           end
 
           ##
-          # Retrieve a single page of RecordingInstance records from the API.
+          # Retrieve a single page of FlowInstance records from the API.
           # Request is executed immediately.
-          # @param [recording.Status] status The status
-          # @param [String] source_sid The source_sid
-          # @param [String] grouping_sid The grouping_sid
-          # @param [Time] date_created_after The date_created_after
-          # @param [Time] date_created_before The date_created_before
           # @param [String] page_token PageToken provided by the API
           # @param [Integer] page_number Page Number, this value is simply for client state
           # @param [Integer] page_size Number of records to return, defaults to 50
-          # @return [Page] Page of RecordingInstance
-          def page(status: :unset, source_sid: :unset, grouping_sid: :unset, date_created_after: :unset, date_created_before: :unset, page_token: :unset, page_number: :unset, page_size: :unset)
+          # @return [Page] Page of FlowInstance
+          def page(page_token: :unset, page_number: :unset, page_size: :unset)
             params = Twilio::Values.of({
-                'Status' => status,
-                'SourceSid' => source_sid,
-                'GroupingSid' => Twilio.serialize_list(grouping_sid) { |e| e },
-                'DateCreatedAfter' => Twilio.serialize_iso8601_datetime(date_created_after),
-                'DateCreatedBefore' => Twilio.serialize_iso8601_datetime(date_created_before),
                 'PageToken' => page_token,
                 'Page' => page_number,
                 'PageSize' => page_size,
@@ -122,36 +89,38 @@ module Twilio
                 @uri,
                 params
             )
-            RecordingPage.new(@version, response, @solution)
+            FlowPage.new(@version, response, @solution)
           end
 
           ##
-          # Retrieve a single page of RecordingInstance records from the API.
+          # Retrieve a single page of FlowInstance records from the API.
           # Request is executed immediately.
           # @param [String] target_url API-generated URL for the requested results page
-          # @return [Page] Page of RecordingInstance
+          # @return [Page] Page of FlowInstance
           def get_page(target_url)
             response = @version.domain.request(
                 'GET',
                 target_url
             )
-            RecordingPage.new(@version, response, @solution)
+            FlowPage.new(@version, response, @solution)
           end
 
           ##
           # Provide a user friendly representation
           def to_s
-            '#<Twilio.Video.V1.RecordingList>'
+            '#<Twilio.Studio.V1.FlowList>'
           end
         end
 
-        class RecordingPage < Page
+        ##
+        # PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
+        class FlowPage < Page
           ##
-          # Initialize the RecordingPage
+          # Initialize the FlowPage
           # @param [Version] version Version that contains the resource
           # @param [Response] response Response from the API
           # @param [Hash] solution Path solution for the resource
-          # @return [RecordingPage] RecordingPage
+          # @return [FlowPage] FlowPage
           def initialize(version, response, solution)
             super(version, response)
 
@@ -160,37 +129,42 @@ module Twilio
           end
 
           ##
-          # Build an instance of RecordingInstance
+          # Build an instance of FlowInstance
           # @param [Hash] payload Payload response from the API
-          # @return [RecordingInstance] RecordingInstance
+          # @return [FlowInstance] FlowInstance
           def get_instance(payload)
-            RecordingInstance.new(@version, payload, )
+            FlowInstance.new(@version, payload, )
           end
 
           ##
           # Provide a user friendly representation
           def to_s
-            '<Twilio.Video.V1.RecordingPage>'
+            '<Twilio.Studio.V1.FlowPage>'
           end
         end
 
-        class RecordingContext < InstanceContext
+        ##
+        # PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
+        class FlowContext < InstanceContext
           ##
-          # Initialize the RecordingContext
+          # Initialize the FlowContext
           # @param [Version] version Version that contains the resource
           # @param [String] sid The sid
-          # @return [RecordingContext] RecordingContext
+          # @return [FlowContext] FlowContext
           def initialize(version, sid)
             super(version)
 
             # Path Solution
             @solution = {sid: sid, }
-            @uri = "/Recordings/#{@solution[:sid]}"
+            @uri = "/Flows/#{@solution[:sid]}"
+
+            # Dependents
+            @engagements = nil
           end
 
           ##
-          # Fetch a RecordingInstance
-          # @return [RecordingInstance] Fetched RecordingInstance
+          # Fetch a FlowInstance
+          # @return [FlowInstance] Fetched FlowInstance
           def fetch
             params = Twilio::Values.of({})
 
@@ -200,49 +174,64 @@ module Twilio
                 params,
             )
 
-            RecordingInstance.new(@version, payload, sid: @solution[:sid], )
+            FlowInstance.new(@version, payload, sid: @solution[:sid], )
           end
 
           ##
-          # Deletes the RecordingInstance
+          # Deletes the FlowInstance
           # @return [Boolean] true if delete succeeds, true otherwise
           def delete
             @version.delete('delete', @uri)
           end
 
           ##
+          # Access the engagements
+          # @return [EngagementList]
+          # @return [EngagementContext] if sid was passed.
+          def engagements(sid=:unset)
+            raise ArgumentError, 'sid cannot be nil' if sid.nil?
+
+            if sid != :unset
+              return EngagementContext.new(@version, @solution[:sid], sid, )
+            end
+
+            unless @engagements
+              @engagements = EngagementList.new(@version, flow_sid: @solution[:sid], )
+            end
+
+            @engagements
+          end
+
+          ##
           # Provide a user friendly representation
           def to_s
             context = @solution.map {|k, v| "#{k}: #{v}"}.join(',')
-            "#<Twilio.Video.V1.RecordingContext #{context}>"
+            "#<Twilio.Studio.V1.FlowContext #{context}>"
           end
         end
 
-        class RecordingInstance < InstanceResource
+        ##
+        # PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
+        class FlowInstance < InstanceResource
           ##
-          # Initialize the RecordingInstance
+          # Initialize the FlowInstance
           # @param [Version] version Version that contains the resource
           # @param [Hash] payload payload that contains response from Twilio
           # @param [String] sid The sid
-          # @return [RecordingInstance] RecordingInstance
+          # @return [FlowInstance] FlowInstance
           def initialize(version, payload, sid: nil)
             super(version)
 
             # Marshaled Properties
             @properties = {
-                'account_sid' => payload['account_sid'],
-                'status' => payload['status'],
-                'date_created' => Twilio.deserialize_iso8601_datetime(payload['date_created']),
                 'sid' => payload['sid'],
-                'source_sid' => payload['source_sid'],
-                'size' => payload['size'] == nil ? payload['size'] : payload['size'].to_i,
+                'account_sid' => payload['account_sid'],
+                'friendly_name' => payload['friendly_name'],
+                'status' => payload['status'],
+                'version' => payload['version'].to_i,
+                'date_created' => Twilio.deserialize_iso8601_datetime(payload['date_created']),
+                'date_updated' => Twilio.deserialize_iso8601_datetime(payload['date_updated']),
                 'url' => payload['url'],
-                'type' => payload['type'],
-                'duration' => payload['duration'] == nil ? payload['duration'] : payload['duration'].to_i,
-                'container_format' => payload['container_format'],
-                'codec' => payload['codec'],
-                'grouping_sids' => payload['grouping_sids'],
-                'track_name' => payload['track_name'],
                 'links' => payload['links'],
             }
 
@@ -254,30 +243,12 @@ module Twilio
           ##
           # Generate an instance context for the instance, the context is capable of
           # performing various actions.  All instance actions are proxied to the context
-          # @return [RecordingContext] RecordingContext for this RecordingInstance
+          # @return [FlowContext] FlowContext for this FlowInstance
           def context
             unless @instance_context
-              @instance_context = RecordingContext.new(@version, @params['sid'], )
+              @instance_context = FlowContext.new(@version, @params['sid'], )
             end
             @instance_context
-          end
-
-          ##
-          # @return [String] The account_sid
-          def account_sid
-            @properties['account_sid']
-          end
-
-          ##
-          # @return [recording.Status] The status
-          def status
-            @properties['status']
-          end
-
-          ##
-          # @return [Time] The date_created
-          def date_created
-            @properties['date_created']
           end
 
           ##
@@ -287,15 +258,39 @@ module Twilio
           end
 
           ##
-          # @return [String] The source_sid
-          def source_sid
-            @properties['source_sid']
+          # @return [String] The account_sid
+          def account_sid
+            @properties['account_sid']
           end
 
           ##
-          # @return [String] The size
-          def size
-            @properties['size']
+          # @return [String] The friendly_name
+          def friendly_name
+            @properties['friendly_name']
+          end
+
+          ##
+          # @return [flow.Status] The status
+          def status
+            @properties['status']
+          end
+
+          ##
+          # @return [String] The version
+          def version
+            @properties['version']
+          end
+
+          ##
+          # @return [Time] The date_created
+          def date_created
+            @properties['date_created']
+          end
+
+          ##
+          # @return [Time] The date_updated
+          def date_updated
+            @properties['date_updated']
           end
 
           ##
@@ -305,73 +300,44 @@ module Twilio
           end
 
           ##
-          # @return [recording.Type] The type
-          def type
-            @properties['type']
-          end
-
-          ##
-          # @return [String] The duration
-          def duration
-            @properties['duration']
-          end
-
-          ##
-          # @return [recording.Format] The container_format
-          def container_format
-            @properties['container_format']
-          end
-
-          ##
-          # @return [recording.Codec] The codec
-          def codec
-            @properties['codec']
-          end
-
-          ##
-          # @return [Hash] The grouping_sids
-          def grouping_sids
-            @properties['grouping_sids']
-          end
-
-          ##
-          # @return [String] The track_name
-          def track_name
-            @properties['track_name']
-          end
-
-          ##
           # @return [String] The links
           def links
             @properties['links']
           end
 
           ##
-          # Fetch a RecordingInstance
-          # @return [RecordingInstance] Fetched RecordingInstance
+          # Fetch a FlowInstance
+          # @return [FlowInstance] Fetched FlowInstance
           def fetch
             context.fetch
           end
 
           ##
-          # Deletes the RecordingInstance
+          # Deletes the FlowInstance
           # @return [Boolean] true if delete succeeds, true otherwise
           def delete
             context.delete
           end
 
           ##
+          # Access the engagements
+          # @return [engagements] engagements
+          def engagements
+            context.engagements
+          end
+
+          ##
           # Provide a user friendly representation
           def to_s
             values = @params.map{|k, v| "#{k}: #{v}"}.join(" ")
-            "<Twilio.Video.V1.RecordingInstance #{values}>"
+            "<Twilio.Studio.V1.FlowInstance #{values}>"
           end
 
           ##
           # Provide a detailed, user friendly representation
           def inspect
             values = @properties.map{|k, v| "#{k}: #{v}"}.join(" ")
-            "<Twilio.Video.V1.RecordingInstance #{values}>"
+            "<Twilio.Studio.V1.FlowInstance #{values}>"
           end
         end
       end