+++ lib/twilio-ruby/rest/video.rb
@@ -28,6 +28,14 @@ module Twilio
 
       ##
       # @param [String] sid The sid
+      # @return [Twilio::REST::Video::V1::CompositionInstance] if sid was passed.
+      # @return [Twilio::REST::Video::V1::CompositionList]
+      def compositions(sid=:unset)
+        self.v1.compositions(sid)
+      end
+
+      ##
+      # @param [String] sid The sid
       # @return [Twilio::REST::Video::V1::RecordingInstance] if sid was passed.
       # @return [Twilio::REST::Video::V1::RecordingList]
       def recordings(sid=:unset)