+++ spec/integration/video/v1/recording_spec.rb
@@ -38,6 +38,7 @@ describe 'Recording' do
           "duration": 0,
           "container_format": "mka",
           "codec": "OPUS",
+          "track_name": "A name",
           "grouping_sids": {
               "room_sid": "RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
           },
@@ -111,6 +112,7 @@ describe 'Recording' do
                   "duration": 10,
                   "container_format": "mka",
                   "codec": "OPUS",
+                  "track_name": "A name",
                   "grouping_sids": {
                       "room_sid": "RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                       "participant_sid": "PAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"